backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
paramiko==3.5.0
ldap3==2.9.1
hvac==2.3.0
redis==5.2.0
cryptography==43.0.3
python-dotenv==1.0.1
jinja2==3.1.4
celery[redis]==5.4.0
celery[redis]==5.4.0
bcrypt==4.0.1
